instance: Cleanup instance type
